feat: add argument validation to component
