+++ styles/components/_buttonRed.scss
@@ -1,22 +1,34 @@
 .button {
-    font-family: $font-primary;
-    border-radius: 1rem;
-    background-color: $color-primary;
-    color: $text-color-primary;
-    border: 0;
-    font-weight: 500;
-    font-size: 2.5rem;
-    padding: 1rem;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    position: relative;
-    min-width: 400px;
-    line-height: 1;
-    
-    &--red {
-        background-color: #ff0000;
-        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25), 0px 4px 30px #FF0000;
-        opacity: .8;
-        color: #fff;
-        text-transform: uppercase;
-    }
+  font-family: $font-primary;
+  border-radius: 1rem;
+  background-color: $color-primary;
+  color: $text-color-primary;
+  border: 0;
+  font-weight: 500;
+  font-size: 2.5rem;
+  padding: 1rem;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  position: relative;
+  min-width: 400px;
+  line-height: 1;
+
+  &--red {
+    background-color: #ff0000;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25), 0px 4px 30px #ff0000;
+    opacity: 0.8;
+    color: #fff;
+    text-transform: uppercase;
+  }
+  &:hover {
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25),
+      inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+  }
+
+  &:focus {
+    outline: none;
+  }
+
+  &:active {
+    transform: scale(0.99);
+  }
 }
